Làm lại toàn bộ, vịtrí thành trạng thái toa tàu
close #5, close #7
